fix(voice-identity): honor core contract layering
